Small changes to variant formatter
